Added groups for tables
Groups can be considered when calling create and drop function so that the tables in a specific group can be controlled separately from everything else. Helps with cogs managing their own tables. Also added a connection pool owned by the DatabaseManager so that connections aren't having to be repeatedly opened and closed.